improve types.
