+++ generators/gradle/types.d.ts
@@ -1,17 +1,16 @@
 import type { RequireOneOrNone } from 'type-fest';
 
 import type { HandleCommandTypes } from '../../lib/command/types.ts';
+import type { Source as BaseApplicationSource } from '../base-simple-application/types.ts';
 import type {
   Application as BuildToolApplication,
   Config as BuildToolConfig,
   Options as BuildToolOptions,
-  Source as BuildToolSource,
 } from '../java/generators/build-tool/types.d.ts';
 import type {
-  Application as JavaBootstrapApplication,
-  Config as JavaBootstrapConfig,
-  Options as JavaBootstrapOptions,
-  Source as JavaBootstrapSource,
+  Application as JavaSimpleApplicationApplication,
+  Config as JavaSimpleApplicationConfig,
+  Options as JavaSimpleApplicationOptions,
 } from '../java-simple-application/types.ts';
 
 import type GradleCommand from './command.ts';
@@ -51,38 +50,37 @@ export type GradleCatalogNeedleOptions = { gradleVersionCatalogFile?: string };
 
 export type GradleNeedleOptions = GradleFileNeedleOptions & GradleCatalogNeedleOptions;
 
-export type Source = JavaBootstrapSource &
-  BuildToolSource & {
-    _gradleDependencies?: GradleDependency[];
-    applyFromGradle?(script: GradleScript): void;
-    addGradleDependency?(dependency: GradleDependency, options?: GradleFileNeedleOptions): void;
-    addGradleDependencies?(dependency: GradleDependency[], options?: GradleFileNeedleOptions): void;
-    addGradlePlugin?(plugin: GradlePlugin): void;
-    addGradlePluginManagement?(pluginManagement: GradlePlugin): void;
-    addGradleProperty?(property: GradleProperty & GradleComment): void;
-    addGradleMavenRepository?(repository: GradleRepository): void;
-    addGradleBuildSrcDependency?(dependency: GradleDependency): void;
+export type Source = BaseApplicationSource & {
+  _gradleDependencies?: GradleDependency[];
+  applyFromGradle?(script: GradleScript): void;
+  addGradleDependency?(dependency: GradleDependency, options?: GradleFileNeedleOptions): void;
+  addGradleDependencies?(dependency: GradleDependency[], options?: GradleFileNeedleOptions): void;
+  addGradlePlugin?(plugin: GradlePlugin): void;
+  addGradlePluginManagement?(pluginManagement: GradlePlugin): void;
+  addGradleProperty?(property: GradleProperty & GradleComment): void;
+  addGradleMavenRepository?(repository: GradleRepository): void;
+  addGradleBuildSrcDependency?(dependency: GradleDependency): void;
 
-    addGradleDependencyCatalogVersion?(catalogVersion: GradleTomlVersion, options?: GradleCatalogNeedleOptions): void;
-    addGradleDependencyCatalogVersions?(catalogVersion: GradleTomlVersion[], options?: GradleCatalogNeedleOptions): void;
-    addGradleDependencyCatalogLibrary?(catalogVersion: GradleLibrary, options?: GradleNeedleOptions): void;
-    addGradleDependencyCatalogLibraries?(catalogVersion: GradleLibrary[], options?: GradleNeedleOptions): void;
-    addGradleDependencyCatalogPlugin?(catalogVersion: GradleTomlPlugin): void;
-    addGradleDependencyCatalogPlugins?(catalogVersion: GradleTomlPlugin[]): void;
+  addGradleDependencyCatalogVersion?(catalogVersion: GradleTomlVersion, options?: GradleCatalogNeedleOptions): void;
+  addGradleDependencyCatalogVersions?(catalogVersion: GradleTomlVersion[], options?: GradleCatalogNeedleOptions): void;
+  addGradleDependencyCatalogLibrary?(catalogVersion: GradleLibrary, options?: GradleNeedleOptions): void;
+  addGradleDependencyCatalogLibraries?(catalogVersion: GradleLibrary[], options?: GradleNeedleOptions): void;
+  addGradleDependencyCatalogPlugin?(catalogVersion: GradleTomlPlugin): void;
+  addGradleDependencyCatalogPlugins?(catalogVersion: GradleTomlPlugin[]): void;
 
-    addGradleBuildSrcDependencyCatalogVersion?(catalogVersion: GradleTomlVersion): void;
-    addGradleBuildSrcDependencyCatalogVersions?(catalogVersion: GradleTomlVersion[]): void;
-    addGradleBuildSrcDependencyCatalogLibraries?(catalogVersion: GradleLibrary[]): void;
-  };
+  addGradleBuildSrcDependencyCatalogVersion?(catalogVersion: GradleTomlVersion): void;
+  addGradleBuildSrcDependencyCatalogVersions?(catalogVersion: GradleTomlVersion[]): void;
+  addGradleBuildSrcDependencyCatalogLibraries?(catalogVersion: GradleLibrary[]): void;
+};
 
 type Command = HandleCommandTypes<typeof GradleCommand>;
 
-export type Config = Command['Config'] & JavaBootstrapConfig & BuildToolConfig;
+export type Config = Command['Config'] & JavaSimpleApplicationConfig & BuildToolConfig;
 
-export type Options = Command['Options'] & JavaBootstrapOptions & BuildToolOptions;
+export type Options = Command['Options'] & JavaSimpleApplicationOptions & BuildToolOptions;
 
 export type Application = Command['Application'] &
-  JavaBootstrapApplication &
+  JavaSimpleApplicationApplication &
   BuildToolApplication & {
     gradleVersion?: string;
     gradleBuildSrc?: string;
