+++ generators/java/types.d.ts
@@ -1,5 +1,3 @@
-import type { RequireOneOrNone } from 'type-fest';
-
 import type { ExportGeneratorOptionsFromCommand, ExportStoragePropertiesFromCommand } from '../../lib/command/index.ts';
 import type {
   Application as BaseApplicationApplication,
@@ -10,16 +8,14 @@ import type {
   Relationship as BaseApplicationRelationship,
   Source as BaseApplicationSource,
 } from '../base-application/types.d.ts';
-import type { EditFileCallback } from '../base-core/api.ts';
-import type { PropertiesFileKeyUpdate } from '../base-core/support/index.ts';
-import type { Application as GradleApplication, GradleNeedleOptions, Source as GradleSource } from '../gradle/types.ts';
+import type { Application as GradleApplication } from '../gradle/types.ts';
 import type {
   Application as JavaSimpleApplicationApplication,
   Config as JavaSimpleApplicationConfig,
   Options as JavaSimpleApplicationOptions,
+  Source as JavaSimpleApplicationSource,
 } from '../java-simple-application/types.ts';
 import type { Application as LanguagesApplication } from '../languages/types.ts';
-import type { MavenDefinition, Source as MavenSource } from '../maven/types.ts';
 
 import type { JavaAddedApplicationProperties } from './application.ts';
 import type {
@@ -28,7 +24,18 @@ import type {
   Options as BuildToolOptions,
 } from './generators/build-tool/types.ts';
 import type GraalvmCommand from './generators/graalvm/command.ts';
-import type { JavaAnnotation } from './support/add-java-annotation.ts';
+
+export type {
+  ConditionalJavaDefinition,
+  JavaArtifact,
+  JavaArtifactType,
+  JavaArtifactVersion,
+  JavaDefinition,
+  JavaDependency,
+  JavaDependencyVersion,
+  JavaNeedleOptions,
+  SpringBean,
+} from '../java-simple-application/types.ts';
 
 type Property = {
   propertyJavaFilterName?: string;
@@ -101,37 +108,6 @@ export interface Entity<F extends Field = Field, R extends Relationship = Relati
   relationshipsContainOtherSideIgnore?: boolean;
 }
 
-export type JavaDependencyVersion = {
-  name: string;
-  version: string;
-};
-
-export type JavaArtifactType = {
-  type?: 'jar' | 'pom';
-  scope?: 'compile' | 'provided' | 'runtime' | 'test' | 'system' | 'import' | 'annotationProcessor';
-};
-
-export type JavaArtifact = {
-  groupId: string;
-  artifactId: string;
-  classifier?: string;
-} & JavaArtifactType;
-
-export type JavaArtifactVersion = RequireOneOrNone<{ version?: string; versionRef?: string }, 'version' | 'versionRef'>;
-
-export type JavaDependency = JavaArtifact &
-  JavaArtifactVersion & {
-    exclusions?: JavaArtifact[];
-  };
-
-export type JavaDefinition = {
-  versions?: JavaDependencyVersion[];
-  dependencies?: JavaDependency[];
-  mavenDefinition?: MavenDefinition;
-};
-
-export type JavaNeedleOptions = GradleNeedleOptions;
-
 export type Config = BaseApplicationConfig &
   JavaSimpleApplicationConfig &
   BuildToolConfig &
@@ -160,71 +136,4 @@ export type Application<E extends BaseApplicationEntity<BaseApplicationField, Ba
     LanguagesApplication &
     DatabaseApplication & {};
 
-export type ConditionalJavaDefinition = JavaDefinition & { condition?: boolean };
-
-export type SpringBean = { package: string; beanClass: string; beanName: string };
-
-export type Source = BaseApplicationSource &
-  MavenSource &
-  GradleSource & {
-    /**
-     * Add a JavaDefinition to the application.
-     * A version requires a valid version otherwise it will be ignored.
-     * A dependency with versionRef requires a valid referenced version at `versions` otherwise it will be ignored.
-     */
-    addJavaDefinition?(definition: JavaDefinition, options?: JavaNeedleOptions): void;
-    addJavaDefinitions?(
-      optionsOrDefinition: JavaNeedleOptions | ConditionalJavaDefinition,
-      ...definitions: ConditionalJavaDefinition[]
-    ): void;
-    addJavaDependencies?(dependency: JavaDependency[], options?: JavaNeedleOptions): void;
-    hasJavaProperty?(propertyName: string): boolean;
-    hasJavaManagedProperty?(propertyName: string): boolean;
-    addMainLog?({ name, level }: { name: string; level: string }): void;
-    addTestLog?({ name, level }: { name: string; level: string }): void;
-
-    editJUnitPlatformProperties?(properties: PropertiesFileKeyUpdate[]): void;
-    /**
-     * Edit a Java file by adding static imports, imports and annotations.
-     * Callbacks are passed to the editFile method.
-     */
-    editJavaFile?: (
-      file: string,
-      options: {
-        staticImports?: string[];
-        imports?: string[];
-        annotations?: JavaAnnotation[];
-        /**
-         * Constructor parameters to add to the class.
-         */
-        constructorParams?: string[];
-        /**
-         * Fields to add to the class.
-         * Requires a valid constructor.
-         */
-        fields?: string[];
-        /**
-         * Spring beans to add to the class.
-         */
-        springBeans?: SpringBean[];
-      },
-      ...editFileCallback: EditFileCallback[]
-    ) => void;
-    /**
-     * Add enum values to a Java enum.
-     *
-     * @example
-     * ```js
-     * addItemsToJavaEnumFile('src/main/java/my/package/MyEnum.java', {
-     *   enumValues: ['VALUE1', 'VALUE2'],
-     * });
-     * ```
-     */
-    addItemsToJavaEnumFile?: (
-      file: string,
-      options: {
-        enumName?: string;
-        enumValues: string[];
-      },
-    ) => void;
-  };
+export type Source = BaseApplicationSource & JavaSimpleApplicationSource;
