+++ generators/javascript-simple-application/application.spec.ts
@@ -1,0 +1,40 @@
+/**
+ * Copyright 2013-2025 the original author or authors from the JHipster project.
+ *
+ * This file is part of the JHipster project, see https://www.jhipster.tech/
+ * for more information.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *      https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+
+import { expect } from 'esmocha';
+
+import { mutateMockedCompleteData, mutateMockedData } from '../../lib/testing/index.ts';
+
+import * as entityData from './application.ts';
+
+describe('application mutation test', () => {
+  for (const [name, data] of Object.entries(entityData)) {
+    it(`expects ${name} to match snapshot`, () => {
+      expect(
+        mutateMockedData(
+          { jhipsterVersion: 'jhipsterVersion', jhipsterPackageJson: 'jhipsterPackageJson', nodeVersion: 'nodeVersion' },
+          data,
+        ),
+      ).toMatchSnapshot();
+    });
+    it(`expects ${name} to don't override existing properties`, () => {
+      expect(Object.keys(mutateMockedCompleteData(data))).toHaveLength(0);
+    });
+  }
+});
