+++ generators/javascript-simple-application/generators/husky/templates/.lintstagedrc.cjs.ejs
@@ -17,5 +17,5 @@
  limitations under the License.
 -%>
 module.exports = {
-  '{<%= prettierFolders %>}*.{<%= prettierExtensions %>}': ['prettier --write'],
+  '{<%= prettierFolders %>}*.{<%= prettierExtensions.join(',') %>}': ['prettier --write'],
 };
